temporal commit

+++ src/main/java/com/mgoportfolio2022api/mgoportfolio2022api/service/dto/AlbumPostDTO.java
@@ -26,7 +26,7 @@ public class AlbumPostDTO {
     private long[] imageIds;
 
     @JsonProperty(value="categoryIds")
-    private Map<Integer, List<Long>> categoryIds;
+    private Map<Long, List<Long>> categoryIds;
 
 
 }
